refactor: split supabaseFunctions into client/tours/annotations modules; extract LoadingSpinner
The Supabase client is now created lazily in supabaseClient.js, so a
deployment with tours and annotations disabled (or missing env vars)
never constructs one; tour functions move next to the tour components
and annotation functions next to the annotation components, making
Supabase reachable only from feature modules ahead of the feature-flag
work. Note: a misconfigured deploy now fails at first tour/annotation
use rather than at page load.

Also extracts the loading-spinner SVG that was copy-pasted in four
components into ui/LoadingSpinner.svelte, and removes a stray .DS_Store.
